Update zfs-auto-snapshot.sh
Changes behaviour to only accept 1 location input. 
This is done for speed reasons on larger pools when you don't want zfs list to search large directories that are known to never get pruned or snapshots created in. 
I.E. 
tank 
  |-->p1 
  |       |--->c1..5000 (children that will never be acted upon)
  |--->p2
  |        |--->c1..5 (children that will be acted upon this invocation) 

Instead of having to search through the 5k children with N snapshots we can skip them and just search the one parent location we know we're looking for and act upon it and its children